requirements.txt
Flask
mysql
opencv-python
numpy
tensorflow
Pillow
transformers
ultralytics
hashlib>=3.0.0
scikit-image
matplotlib